// src/enabledElements.d.ts
export interface vec2 {
    x: number;
    y: number;
}
/**
 * VOI
 *
 * @typedef {Object} VOI
 * @memberof Objects
 * @param {Number} windowWidth - Window Width for display
 * @param {Number} windowCenter - Window Center for display
 */
export interface VOI {
    windowWidth: number;
    windowCenter: number;
}
/**
 * Lookup Table Array
 *
 * @typedef {Object} LUT
 * @memberof Objects
 * @property {Number} firstValueMapped
 * @property {Number} numBitsPerEntry
 * @property {Array} lut
 */
export interface LUT {
    firstValueMapped: number;
    numBitsPerEntry: number;
    lut: number[];
    id?: string;
}
// ...
export interface Image {
    imageId: string;
    minPixelValue: number;
    maxPixelValue: number;
    slope: number;
    intercept: number;
    windowCenter: number;
    windowWidth: number;
    getPixelData: () => any;
    getImageData: () => any;
    getCanvas: () => any;
    getImage: () => any;
    rows: number;
    columns: number;
    height: number;
    width: number;
    color: boolean;
    lut: LUT;
    rgba: boolean;
    columnPixelSpacing: number;
    rowPixelSpacing: number;
    invert: boolean;
    sizeInBytes: number;
    falseColor?: boolean;
    origPixelData?: ArrayLike<number>;
    colormap?: string | Colormap;
    labelmap?: boolean;
    [s: string]: any;
}
export interface Colormap {
    id: string;
    name: string;
    min: number;
    max: number;
    colors: string[];
}
// ...
export interface Viewport {
    scale?: number;
    translation?: vec2;
    voi?: VOI;
    invert?: boolean;
    pixelReplication?: boolean;
    hflip?: boolean;
    vflip?: boolean;
    rotation?: number;
    modalityLUT?: LUT;
    voiLUT?: LUT;
    colormap?: string | Colormap;
    labelmap?: boolean;
    [s: string]: any;
}
// ...
export interface EnabledElementLayer {
    element?: HTMLElement;
    image: Image;
    viewport: Viewport | undefined;
    canvas?: HTMLCanvasElement;
    options: any;
    invalid?: boolean;
    needsRedraw?: boolean;
    [s: string]: any;
}